+++ docs/src/content/docs/changelog.md
@@ -8,6 +8,9 @@ Requires libvips v8.16.0
 
 ### v0.34.0 - TBD
 
+* Breaking: Support array of input images to be joined or animated.
+  [#1580](https://github.com/lovell/sharp/issues/1580)
+
 * Breaking: Support `info.size` on wide-character systems via upgrade to C++17.
   [#3943](https://github.com/lovell/sharp/issues/3943)